Moved BottomSheetDialogFragment class
